Move Components used in History page to components/History
